yum, net graphs
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky@28 7f574dfc-610e-0410-a909-a81674777703

+++ common.c
@@ -146,7 +146,7 @@ void format_seconds_short(char *buf, unsigned int n, long t)
 static double last_meminfo_update;
 static double last_fs_update;
 
-unsigned int need_mask;
+unsigned long long need_mask;
 
 void update_stuff()
 {
